refactor(desktop): move domain types to slice types.ts files
store.ts should be composition-only per agents.md. moved BootPhase,
SystemAlert, and SessionNudge to their owning slice types.ts; store.ts
re-exports them so external consumers are unaffected.
