+++ apps/desktop/src/store/slices/boot/types.ts
@@ -1,1 +1,11 @@
 export type { SetFn, GetFn } from '../../slice-types';
+
+export type BootPhase =
+  | 'pending'
+  | 'migrating'
+  | 'loading-settings'
+  | 'detecting-cli'
+  | 'loading-workspaces'
+  | 'restoring-session'
+  | 'ready'
+  | 'error';
